created general MenuItem model class
